+++ src/main/java/be/robbevanherck/javafraggenescan/repositories/InputFileRepository.java
@@ -5,7 +5,6 @@ import be.robbevanherck.javafraggenescan.entities.*;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.math.BigDecimal;
 import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.Map;
@@ -15,10 +14,10 @@ import java.util.Scanner;
  * Repository for the data in the input HMM file
  */
 public class InputFileRepository {
-    private final Map<HMMInnerTransition, BigDecimal> innerTransitions;
-    private final Map<HMMOuterTransition, BigDecimal> outerTransitions;
-    private final Map<Pair<AminoAcid>, BigDecimal> matchInsertEmissions;
-    private final Map<Pair<AminoAcid>, BigDecimal> insertInsertEmissions;
+    private final Map<HMMInnerTransition, Double> innerTransitions;
+    private final Map<HMMOuterTransition, Double> outerTransitions;
+    private final Map<Pair<AminoAcid>, Double> matchInsertEmissions;
+    private final Map<Pair<AminoAcid>, Double> insertInsertEmissions;
     private final Map<HMMState, PathProbability> initialProbabilities;
     private static InputFileRepository instance;
 
@@ -37,7 +36,7 @@ public class InputFileRepository {
                 /* Outer Transitions */
                 for (int i = 0; i < HMMOuterTransition.values().length - 1; i++) {
                     String transitionString = s.next();
-                    BigDecimal transitionProbability = BigDecimal.valueOf(s.nextDouble());
+                    double transitionProbability = s.nextDouble();
 
                     HMMOuterTransition transition = HMMOuterTransition.fromString(transitionString);
 
@@ -47,7 +46,7 @@ public class InputFileRepository {
                 /* Inner Transitions */
                 for (int i = 0; i < HMMInnerTransition.values().length - 1; i++) {
                     String transitionString = s.next();
-                    BigDecimal transitionProbability = BigDecimal.valueOf(s.nextDouble());
+                    double transitionProbability = s.nextDouble();
 
                     HMMInnerTransition transition = HMMInnerTransition.fromString(transitionString);
 
@@ -59,7 +58,7 @@ public class InputFileRepository {
                 for (int i = 0; i < 16; i++) {
                     String firstAminoAcidString = s.next();
                     String secondAminoAcidString = s.next();
-                    BigDecimal miEmissionProbability = BigDecimal.valueOf(s.nextDouble());
+                    double miEmissionProbability = s.nextDouble();
 
                     AminoAcid firstAminoAcid = AminoAcid.fromString(firstAminoAcidString);
                     AminoAcid secondAminoAcid = AminoAcid.fromString(secondAminoAcidString);
@@ -73,7 +72,7 @@ public class InputFileRepository {
                 for (int i = 0; i < 16; i++) {
                     String firstAminoAcidString = s.next();
                     String secondAminoAcidString = s.next();
-                    BigDecimal iiEmissionProbability = BigDecimal.valueOf(s.nextDouble());
+                    double iiEmissionProbability = s.nextDouble();
 
                     AminoAcid firstAminoAcid = AminoAcid.fromString(firstAminoAcidString);
                     AminoAcid secondAminoAcid = AminoAcid.fromString(secondAminoAcidString);
@@ -86,7 +85,7 @@ public class InputFileRepository {
                 s.next(); // Skip the header
                 for (int i = 0; i < HMMState.values().length - 1; i++) {
                     String stateString = s.next();
-                    BigDecimal stateProbability = BigDecimal.valueOf(s.nextDouble());
+                    double stateProbability = s.nextDouble();
 
                     HMMState state = HMMState.fromString(stateString);
 
@@ -110,19 +109,19 @@ public class InputFileRepository {
         return instance;
     }
 
-    public Map<HMMOuterTransition, BigDecimal> getOuterTransitions() {
+    public Map<HMMOuterTransition, Double> getOuterTransitions() {
         return outerTransitions;
     }
 
-    public Map<HMMInnerTransition, BigDecimal> getInnerTransitions() {
+    public Map<HMMInnerTransition, Double> getInnerTransitions() {
         return innerTransitions;
     }
 
-    public Map<Pair<AminoAcid>, BigDecimal> getMatchInsertEmissions() {
+    public Map<Pair<AminoAcid>, Double> getMatchInsertEmissions() {
         return matchInsertEmissions;
     }
 
-    public Map<Pair<AminoAcid>, BigDecimal> getInsertInsertEmissions() {
+    public Map<Pair<AminoAcid>, Double> getInsertInsertEmissions() {
         return insertInsertEmissions;
     }
