+++ src/test/java/be/robbevanherck/javafraggenescan/dummies/NonAbstractMatchTransition.java
@@ -4,8 +4,6 @@ import be.robbevanherck.javafraggenescan.entities.*;
 import be.robbevanherck.javafraggenescan.transitions.InsertTransition;
 import be.robbevanherck.javafraggenescan.transitions.MatchTransition;
 
-import java.math.BigDecimal;
-
 public class NonAbstractMatchTransition extends MatchTransition {
     /**
      * Create a new Dummy InsertTransition that isn't abstract.
@@ -21,16 +19,16 @@ public class NonAbstractMatchTransition extends MatchTransition {
 
     @Override
     protected PathProbability getProbabilityFromMatch(HMMParameters parameters, ViterbiStep previous, Triple<AminoAcid> codonEndingAtT) {
-        return new PathProbability(HMMState.MATCH_1, BigDecimal.valueOf(0.01));
+        return new PathProbability(HMMState.MATCH_1, 0.01);
     }
 
     @Override
     protected PathProbability getProbabilityThroughDeletions(HMMParameters parameters, ViterbiStep previous, Triple<AminoAcid> codonEndingAtT) {
-        return new PathProbability(HMMState.MATCH_2, BigDecimal.valueOf(0.02));
+        return new PathProbability(HMMState.MATCH_2, 0.02);
     }
 
     @Override
     protected PathProbability getProbabilityFromStart(ViterbiStep currentStep, Triple<AminoAcid> codonEndingAtT) {
-        return new PathProbability(HMMState.START, BigDecimal.valueOf(0.03));
+        return new PathProbability(HMMState.START, 0.03);
     }
 }
